services/data-processor/requirements.txt
pymongo
kafka-python
python-dotenv
fastapi
uvicorn
